+++ src/OpenTelemetry.Instrumentation.ServiceFabricRemoting/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+* Add `net8.0`, `net10.0`, and `net462` target frameworks.
+  ([#3791](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3791))
+
+* Updated OpenTelemetry core component version(s) to `1.15.0`.
+  ([#3791](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3791))
+
 ## 1.14.0-beta.1
 
 Released 2025-Nov-13