+++ src/OpenTelemetry.Instrumentation.ServiceFabricRemoting/CHANGELOG.md
@@ -2,12 +2,12 @@
 
 ## Unreleased
 
-* Updated OpenTelemetry core component version(s) to `1.15.2`.
-  ([#4080](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4080))
-
 * Ensure that `TransportSettings` configuration is applied to created
   instances of `IServiceRemotingClientFactory`.
   ([#4148](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4148))
+
+* Updated OpenTelemetry core component version(s) to `1.15.3`.
+  ([#4166](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4166))
 
 ## 1.15.0-beta.1